add functions to lock jobs
DANGER WILL ROBINSON

if you hold locks be VERY careful to not call a function that will
lock, as you will be stuck in lock contention.  And that's much akin
to being eaten by a grue.  This is extra tricky now that there are multiple
locks.  So say we all.

darcs-hash:20050803234847-b89eb-f7009cecee3cb0c92d8a607164bc8c2456777e59.gz