add first / last article number to newsgroups and save postsetjob status in a crc named file
I haven't done anything with this but get it to compile.  Might work, might not.  Need
the 2 missing pieces to be able to tell for sure, namely the ability to track new postsets
in subscribed groups, and the master job file / loading of postset jobs

Well, 3.  Also need to be able to regrab headers for postsets that have been save / restored

darcs-hash:20050915020030-55968-ff81b64025042dd7207dd0065a5140bbe81c4103.gz